Move some tests from PhpInternalSourceLocatorTest to ReflectionSourceStubberTest
